Add comment|check authentication
