+++ app/controllers/api/v1/posts_controller.rb
@@ -8,25 +8,8 @@ module Api
         @posts = Post.all
         render json: @posts
       end
-      
-      def comments
-        @comments = @post.comments
-        render json: @comments
-      end
-
-      def add_comment
-        @comment = @post.comments.build(comment_params)
-        @comment.author_id = @user.id
-
-        if @comment.save
-          render json: @comment, status: :created
-        else
-          render json: @comment.errors, status: :unprocessable_entity
-        end
-      end
 
       private
-
       def set_user
         @user = User.find(params[:user_id])
       end
@@ -35,9 +18,6 @@ module Api
         @post = @user.posts.find(params[:id])
       end
 
-      def comment_params
-        params.require(:comment).permit(:text)
-      end
     end
   end
 end
